--- operator/webhooks/config.go
@@ -1,7 +1,0 @@
-package webhooks
-
-const (
-
-	// SecretCertsName is the name of the secret where the webhook certificates are stored.
-	SecretCertsName = "klt-certs"
-)
