+++ operator/webhooks/pod_mutator/pod_mutating_webhook_test.go
@@ -3,7 +3,6 @@ package pod_mutator
 import (
 	"context"
 	"encoding/json"
-	"reflect"
 	"testing"
 
 	"github.com/go-logr/logr"
@@ -14,7 +13,6 @@ import (
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/trace"
 	admissionv1 "k8s.io/api/admission/v1"
-	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -23,79 +21,6 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/webhook/admission"
 )
-
-func TestPodMutatingWebhook_getOwnerReference(t *testing.T) {
-	type fields struct {
-		Client   client.Client
-		Tracer   trace.Tracer
-		decoder  *admission.Decoder
-		Recorder record.EventRecorder
-		Log      logr.Logger
-	}
-	type args struct {
-		resource *metav1.ObjectMeta
-	}
-	tests := []struct {
-		name   string
-		fields fields
-		args   args
-		want   metav1.OwnerReference
-	}{
-		{
-			name: "Test simple return when UID and Kind is set",
-			args: args{
-				resource: &metav1.ObjectMeta{
-					UID: "the-pod-uid",
-					OwnerReferences: []metav1.OwnerReference{
-						{
-							Kind: "ReplicaSet",
-							UID:  "the-replicaset-uid",
-							Name: "some-name",
-						},
-					},
-				},
-			},
-			want: metav1.OwnerReference{
-				UID:  "the-replicaset-uid",
-				Kind: "ReplicaSet",
-				Name: "some-name",
-			},
-		},
-		{
-			name: "Test return is input argument if owner is not found",
-			args: args{
-				resource: &metav1.ObjectMeta{
-					UID: "the-pod-uid",
-					OwnerReferences: []metav1.OwnerReference{
-						{
-							Kind: "SomeNonExistentType",
-							UID:  "the-replicaset-uid",
-						},
-					},
-				},
-			},
-			want: metav1.OwnerReference{
-				UID:  "",
-				Kind: "",
-				Name: "",
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			a := &PodMutatingWebhook{
-				Client:   tt.fields.Client,
-				Tracer:   tt.fields.Tracer,
-				decoder:  tt.fields.decoder,
-				Recorder: tt.fields.Recorder,
-				Log:      tt.fields.Log,
-			}
-			if got := a.getOwnerReference(tt.args.resource); !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("getOwnerReference() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
 
 func TestPodMutatingWebhook_getAppName(t *testing.T) {
 	type fields struct {
@@ -248,674 +173,6 @@ func TestPodMutatingWebhook_getWorkloadName(t *testing.T) {
 			}
 			if got := a.getWorkloadName(tt.args.pod); got != tt.want {
 				t.Errorf("getWorkloadName() = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func Test_getLabelOrAnnotation(t *testing.T) {
-	type args struct {
-		resource            *metav1.ObjectMeta
-		primaryAnnotation   string
-		secondaryAnnotation string
-	}
-	tests := []struct {
-		name  string
-		args  args
-		want  string
-		want1 bool
-	}{
-		{
-			name: "Test if primary annotation is returned from annotations",
-			args: args{
-				resource: &metav1.ObjectMeta{
-					Annotations: map[string]string{
-						apicommon.AppAnnotation: "some-app-name",
-					},
-				},
-				primaryAnnotation:   apicommon.AppAnnotation,
-				secondaryAnnotation: apicommon.K8sRecommendedAppAnnotations,
-			},
-			want:  "some-app-name",
-			want1: true,
-		},
-		{
-			name: "Test if secondary annotation is returned from annotations",
-			args: args{
-				resource: &metav1.ObjectMeta{
-					Annotations: map[string]string{
-						apicommon.K8sRecommendedAppAnnotations: "some-app-name",
-					},
-				},
-				primaryAnnotation:   apicommon.AppAnnotation,
-				secondaryAnnotation: apicommon.K8sRecommendedAppAnnotations,
-			},
-			want:  "some-app-name",
-			want1: true,
-		},
-		{
-			name: "Test if primary annotation is returned from labels",
-			args: args{
-				resource: &metav1.ObjectMeta{
-					Labels: map[string]string{
-						apicommon.AppAnnotation: "some-app-name",
-					},
-				},
-				primaryAnnotation:   apicommon.AppAnnotation,
-				secondaryAnnotation: apicommon.K8sRecommendedAppAnnotations,
-			},
-			want:  "some-app-name",
-			want1: true,
-		},
-		{
-			name: "Test if secondary annotation is returned from labels",
-			args: args{
-				resource: &metav1.ObjectMeta{
-					Labels: map[string]string{
-						apicommon.K8sRecommendedAppAnnotations: "some-app-name",
-					},
-				},
-				primaryAnnotation:   apicommon.AppAnnotation,
-				secondaryAnnotation: apicommon.K8sRecommendedAppAnnotations,
-			},
-			want:  "some-app-name",
-			want1: true,
-		},
-		{
-			name: "Test that empty string is returned when no annotations or labels are found",
-			args: args{
-				resource: &metav1.ObjectMeta{
-					Annotations: map[string]string{
-						"some-other-annotation": "some-app-name",
-					},
-				},
-				primaryAnnotation:   apicommon.AppAnnotation,
-				secondaryAnnotation: apicommon.K8sRecommendedAppAnnotations,
-			},
-			want:  "",
-			want1: false,
-		},
-		{
-			name: "Test that empty string is returned when primary annotation cannot be found and secondary annotation is empty",
-			args: args{
-				resource: &metav1.ObjectMeta{
-					Annotations: map[string]string{
-						"some-other-annotation": "some-app-name",
-					},
-				},
-				primaryAnnotation:   apicommon.AppAnnotation,
-				secondaryAnnotation: "",
-			},
-			want:  "",
-			want1: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, got1 := getLabelOrAnnotation(tt.args.resource, tt.args.primaryAnnotation, tt.args.secondaryAnnotation)
-			if got != tt.want {
-				t.Errorf("getLabelOrAnnotation() got = %v, want %v", got, tt.want)
-			}
-			if got1 != tt.want1 {
-				t.Errorf("getLabelOrAnnotation() got1 = %v, want %v", got1, tt.want1)
-			}
-		})
-	}
-}
-
-func TestPodMutatingWebhook_isPodAnnotated(t *testing.T) {
-	type fields struct {
-		Client   client.Client
-		Tracer   trace.Tracer
-		decoder  *admission.Decoder
-		Recorder record.EventRecorder
-		Log      logr.Logger
-	}
-	type args struct {
-		pod *corev1.Pod
-	}
-	tests := []struct {
-		name      string
-		fields    fields
-		args      args
-		want      bool
-		wantErr   bool
-		wantedPod *corev1.Pod
-	}{
-		{
-			name: "Test error when workload name is too long",
-			args: args{
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Annotations: map[string]string{
-							apicommon.AppAnnotation:      "SOME-APP-NAME-ANNOTATION",
-							apicommon.WorkloadAnnotation: "workload-name-that-is-too-loooooooooooooooooooooooooooooooooooooooooooooooooong",
-						},
-					},
-				},
-			},
-			want:    false,
-			wantErr: true,
-		},
-		{
-			name: "Test return true when pod has workload annotation",
-			args: args{
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Annotations: map[string]string{
-							apicommon.WorkloadAnnotation: "some-workload-name",
-						},
-					},
-				},
-			},
-			want:    true,
-			wantErr: false,
-		},
-		{
-			name: "Test return true and initialize annotations when labels are set",
-			args: args{
-				pod: &corev1.Pod{
-					Spec: corev1.PodSpec{
-						Containers: []corev1.Container{
-							{
-								Image: "some-image:v1",
-							},
-						},
-					},
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							apicommon.WorkloadAnnotation: "some-workload-name",
-						},
-					},
-				},
-			},
-			want:    true,
-			wantErr: false,
-			wantedPod: &corev1.Pod{
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Image: "some-image:v1",
-						},
-					},
-				},
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: map[string]string{
-						apicommon.WorkloadAnnotation: "some-workload-name",
-					},
-					Annotations: map[string]string{
-						apicommon.VersionAnnotation: "v1",
-					},
-				},
-			},
-		},
-		{
-			name: "Test return false when annotations and labels are not set",
-			args: args{
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Labels: map[string]string{
-							"some-other-label": "some-value",
-						},
-					},
-				},
-			},
-			want:    false,
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			a := &PodMutatingWebhook{
-				Client:   tt.fields.Client,
-				Tracer:   tt.fields.Tracer,
-				decoder:  tt.fields.decoder,
-				Recorder: tt.fields.Recorder,
-				Log:      tt.fields.Log,
-			}
-			got, err := a.isPodAnnotated(tt.args.pod)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("isPodAnnotated() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("isPodAnnotated() got = %v, want %v", got, tt.want)
-			}
-			if tt.wantedPod != nil {
-				require.Equal(t, tt.wantedPod, tt.args.pod)
-			}
-		})
-	}
-}
-
-func TestPodMutatingWebhook_copyAnnotationsIfParentAnnotated(t *testing.T) {
-	testNamespace := "test-namespace"
-	rsUidWithDpOwner := types.UID("this-is-the-replicaset-with-dp-owner")
-	rsUidWithNoOwner := types.UID("this-is-the-replicaset-with-no-owner")
-	testStsUid := types.UID("this-is-the-stateful-set-uid")
-	tstStsName := "test-stateful-set"
-	testDsUid := types.UID("this-is-the-daemon-set-uid")
-	testDsName := "test-daemon-set"
-
-	rsWithDpOwner := &appsv1.ReplicaSet{
-		TypeMeta: metav1.TypeMeta{
-			Kind: "ReplicaSet",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-replicaset1",
-			UID:       rsUidWithDpOwner,
-			Namespace: testNamespace,
-			OwnerReferences: []metav1.OwnerReference{
-				{
-					Kind: "Deployment",
-					Name: "this-is-the-deployment",
-					UID:  "this-is-the-deployment-uid",
-				},
-			},
-		},
-	}
-	// TODO: fix tests where an RS has a STS or DS as owner. they should not have a RS in between
-	rsWithNoOwner := &appsv1.ReplicaSet{
-		TypeMeta: metav1.TypeMeta{
-			Kind: "ReplicaSet",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-replicaset4",
-			UID:       rsUidWithNoOwner,
-			Namespace: testNamespace,
-		},
-	}
-	testDp := &appsv1.Deployment{
-		TypeMeta: metav1.TypeMeta{
-			Kind: "Deployment",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test-deployment",
-			UID:       "this-is-the-deployment-uid",
-			Namespace: testNamespace,
-		},
-	}
-	testSts := &appsv1.StatefulSet{
-		TypeMeta: metav1.TypeMeta{
-			Kind: "StatefulSet",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      tstStsName,
-			UID:       testStsUid,
-			Namespace: testNamespace,
-		},
-	}
-	testDs := &appsv1.DaemonSet{
-		TypeMeta: metav1.TypeMeta{
-			Kind: "DaemonSet",
-		},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      testDsName,
-			UID:       testDsUid,
-			Namespace: testNamespace,
-		},
-	}
-
-	fakeClient := fakeclient.NewClient(rsWithDpOwner, rsWithNoOwner, testDp, testSts, testDs)
-
-	type fields struct {
-		Client   client.Client
-		Tracer   trace.Tracer
-		decoder  *admission.Decoder
-		Recorder record.EventRecorder
-		Log      logr.Logger
-	}
-	type args struct {
-		ctx context.Context
-		req *admission.Request
-		pod *corev1.Pod
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		want    bool
-		wantErr bool
-	}{
-		{
-			name: "Test that nothing happens if owner UID is pod UID",
-			fields: fields{
-				Log:    testr.New(t),
-				Client: fakeClient,
-			},
-			args: args{
-				ctx: context.TODO(),
-				req: &admission.Request{
-					AdmissionRequest: admissionv1.AdmissionRequest{
-						Namespace: testNamespace,
-					},
-				},
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						UID: "some-uid",
-					},
-				},
-			},
-			want:    false,
-			wantErr: false,
-		},
-		{
-			name: "Test fetching of replicaset owner of pod and deployment owner of replicaset",
-			fields: fields{
-				Log:    testr.New(t),
-				Client: fakeClient,
-			},
-			args: args{
-				ctx: context.TODO(),
-				req: &admission.Request{
-					AdmissionRequest: admissionv1.AdmissionRequest{
-						Namespace: testNamespace,
-					},
-				},
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						UID: "this-is-the-pod-uid",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								Name: rsWithDpOwner.Name,
-								UID:  rsUidWithDpOwner,
-								Kind: "ReplicaSet",
-							},
-						},
-					},
-				},
-			},
-			want:    false,
-			wantErr: false,
-		},
-		{
-			name: "Test fetching of statefulset owner of pod",
-			fields: fields{
-				Log:    testr.New(t),
-				Client: fakeClient,
-			},
-			args: args{
-				ctx: context.TODO(),
-				req: &admission.Request{
-					AdmissionRequest: admissionv1.AdmissionRequest{
-						Namespace: testNamespace,
-					},
-				},
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						UID: "this-is-the-pod-uid",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								Name: testSts.Name,
-								UID:  testSts.UID,
-								Kind: testSts.Kind,
-							},
-						},
-					},
-				},
-			},
-			want:    false,
-			wantErr: false,
-		},
-		{
-			name: "Test fetching of daemonset owner of pod",
-			fields: fields{
-				Log:    testr.New(t),
-				Client: fakeClient,
-			},
-			args: args{
-				ctx: context.TODO(),
-				req: &admission.Request{
-					AdmissionRequest: admissionv1.AdmissionRequest{
-						Namespace: testNamespace,
-					},
-				},
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						UID: "this-is-the-pod-uid",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								Name: testDs.Name,
-								UID:  testDs.UID,
-								Kind: testDs.Kind,
-							},
-						},
-					},
-				},
-			},
-			want:    false,
-			wantErr: false,
-		},
-		{
-			name: "Test that method returns without doing anything when we get a pod with replicaset without owner",
-			fields: fields{
-				Log:    testr.New(t),
-				Client: fakeClient,
-			},
-			args: args{
-				ctx: context.TODO(),
-				req: &admission.Request{
-					AdmissionRequest: admissionv1.AdmissionRequest{
-						Namespace: testNamespace,
-					},
-				},
-				pod: &corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						UID: "this-is-the-pod-uid",
-						OwnerReferences: []metav1.OwnerReference{
-							{
-								Name: rsWithNoOwner.Name,
-								UID:  rsUidWithNoOwner,
-								Kind: "ReplicaSet",
-							},
-						},
-					},
-				},
-			},
-			want:    false,
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			a := &PodMutatingWebhook{
-				Client:   tt.fields.Client,
-				Tracer:   tt.fields.Tracer,
-				decoder:  tt.fields.decoder,
-				Recorder: tt.fields.Recorder,
-				Log:      tt.fields.Log,
-			}
-			got, err := a.copyAnnotationsIfParentAnnotated(tt.args.ctx, tt.args.req, tt.args.pod)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("copyAnnotationsIfParentAnnotated() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("copyAnnotationsIfParentAnnotated() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
-}
-
-func TestPodMutatingWebhook_copyResourceLabelsIfPresent(t *testing.T) {
-	type fields struct {
-		Client   client.Client
-		Tracer   trace.Tracer
-		decoder  *admission.Decoder
-		Recorder record.EventRecorder
-		Log      logr.Logger
-	}
-	type args struct {
-		sourceResource *metav1.ObjectMeta
-		targetPod      *corev1.Pod
-	}
-	tests := []struct {
-		name      string
-		fields    fields
-		args      args
-		want      bool
-		wantErr   bool
-		wantedPod *corev1.Pod
-	}{
-		{
-			name: "Test that annotations get copied from source to target",
-			args: args{
-				sourceResource: &metav1.ObjectMeta{
-					Name: "testSourceObject",
-					Annotations: map[string]string{
-						apicommon.WorkloadAnnotation:                 "some-workload-name",
-						apicommon.AppAnnotation:                      "some-app-name",
-						apicommon.VersionAnnotation:                  "v1.0.0",
-						apicommon.PreDeploymentTaskAnnotation:        "some-pre-deployment-task",
-						apicommon.PostDeploymentTaskAnnotation:       "some-post-deployment-task",
-						apicommon.PreDeploymentEvaluationAnnotation:  "some-pre-deployment-evaluation",
-						apicommon.PostDeploymentEvaluationAnnotation: "some-post-deployment-evaluation",
-					},
-				},
-				targetPod: &corev1.Pod{
-					TypeMeta:   metav1.TypeMeta{},
-					ObjectMeta: metav1.ObjectMeta{},
-					Spec:       corev1.PodSpec{},
-					Status:     corev1.PodStatus{},
-				},
-			},
-			want:    true,
-			wantErr: false,
-			wantedPod: &corev1.Pod{
-				TypeMeta: metav1.TypeMeta{},
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						apicommon.WorkloadAnnotation:                 "some-workload-name",
-						apicommon.AppAnnotation:                      "some-app-name",
-						apicommon.VersionAnnotation:                  "v1.0.0",
-						apicommon.PreDeploymentTaskAnnotation:        "some-pre-deployment-task",
-						apicommon.PostDeploymentTaskAnnotation:       "some-post-deployment-task",
-						apicommon.PreDeploymentEvaluationAnnotation:  "some-pre-deployment-evaluation",
-						apicommon.PostDeploymentEvaluationAnnotation: "some-post-deployment-evaluation",
-					},
-				},
-			},
-		},
-		{
-			name: "Test that source labels get copied to target annotations",
-			args: args{
-				sourceResource: &metav1.ObjectMeta{
-					Name: "testSourceObject",
-					Labels: map[string]string{
-						apicommon.WorkloadAnnotation:                 "some-workload-name",
-						apicommon.AppAnnotation:                      "some-app-name",
-						apicommon.VersionAnnotation:                  "v1.0.0",
-						apicommon.PreDeploymentTaskAnnotation:        "some-pre-deployment-task",
-						apicommon.PostDeploymentTaskAnnotation:       "some-post-deployment-task",
-						apicommon.PreDeploymentEvaluationAnnotation:  "some-pre-deployment-evaluation",
-						apicommon.PostDeploymentEvaluationAnnotation: "some-post-deployment-evaluation",
-					},
-				},
-				targetPod: &corev1.Pod{},
-			},
-			want:    true,
-			wantErr: false,
-			wantedPod: &corev1.Pod{
-				TypeMeta: metav1.TypeMeta{},
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						apicommon.WorkloadAnnotation:                 "some-workload-name",
-						apicommon.AppAnnotation:                      "some-app-name",
-						apicommon.VersionAnnotation:                  "v1.0.0",
-						apicommon.PreDeploymentTaskAnnotation:        "some-pre-deployment-task",
-						apicommon.PostDeploymentTaskAnnotation:       "some-post-deployment-task",
-						apicommon.PreDeploymentEvaluationAnnotation:  "some-pre-deployment-evaluation",
-						apicommon.PostDeploymentEvaluationAnnotation: "some-post-deployment-evaluation",
-					},
-				},
-			},
-		},
-		{
-			name: "Test that version label is generated correctly and rest is copied",
-			args: args{
-				sourceResource: &metav1.ObjectMeta{
-					Name: "testSourceObject",
-					Labels: map[string]string{
-						apicommon.WorkloadAnnotation:                 "some-workload-name",
-						apicommon.AppAnnotation:                      "some-app-name",
-						apicommon.PreDeploymentTaskAnnotation:        "some-pre-deployment-task",
-						apicommon.PostDeploymentTaskAnnotation:       "some-post-deployment-task",
-						apicommon.PreDeploymentEvaluationAnnotation:  "some-pre-deployment-evaluation",
-						apicommon.PostDeploymentEvaluationAnnotation: "some-post-deployment-evaluation",
-					},
-				},
-				targetPod: &corev1.Pod{
-					TypeMeta:   metav1.TypeMeta{},
-					ObjectMeta: metav1.ObjectMeta{},
-					Spec: corev1.PodSpec{
-						Containers: []corev1.Container{
-							{
-								Image: "some-image:v1.0.0",
-							},
-						},
-					},
-					Status: corev1.PodStatus{},
-				},
-			},
-			want:    true,
-			wantErr: false,
-			wantedPod: &corev1.Pod{
-				TypeMeta: metav1.TypeMeta{},
-				ObjectMeta: metav1.ObjectMeta{
-					Annotations: map[string]string{
-						apicommon.WorkloadAnnotation:                 "some-workload-name",
-						apicommon.AppAnnotation:                      "some-app-name",
-						apicommon.VersionAnnotation:                  "v1.0.0",
-						apicommon.PreDeploymentTaskAnnotation:        "some-pre-deployment-task",
-						apicommon.PostDeploymentTaskAnnotation:       "some-post-deployment-task",
-						apicommon.PreDeploymentEvaluationAnnotation:  "some-pre-deployment-evaluation",
-						apicommon.PostDeploymentEvaluationAnnotation: "some-post-deployment-evaluation",
-					},
-				},
-				Spec: corev1.PodSpec{
-					Containers: []corev1.Container{
-						{
-							Image: "some-image:v1.0.0",
-						},
-					},
-				},
-				Status: corev1.PodStatus{},
-			},
-		},
-		{
-			name: "Test that error is return with too long workload name",
-			args: args{
-				sourceResource: &metav1.ObjectMeta{
-					Name: "testSourceObject",
-					Labels: map[string]string{
-						apicommon.WorkloadAnnotation: "some-workload-name-that-is-very-looooooooooooooooooooooong",
-					},
-				},
-				targetPod: &corev1.Pod{},
-			},
-			want:    false,
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			a := &PodMutatingWebhook{
-				Client:   tt.fields.Client,
-				Tracer:   tt.fields.Tracer,
-				decoder:  tt.fields.decoder,
-				Recorder: tt.fields.Recorder,
-				Log:      tt.fields.Log,
-			}
-			got, err := a.copyResourceLabelsIfPresent(tt.args.sourceResource, tt.args.targetPod)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("copyResourceLabelsIfPresent() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("copyResourceLabelsIfPresent() got = %v, want %v", got, tt.want)
-			}
-			if tt.wantedPod != nil {
-				require.Equal(t, tt.wantedPod, tt.args.targetPod)
 			}
 		})
 	}
